[FIX] purchase_advance_payment: Relate correctly the payments to orders
There was a problem when registering payments for the moves, meaning the
cases where in the tree view of Moves, registering payments for multiple
moves wouldn't relate these to the related orders.

Also another fix is to correctly change the payment status for the order
when it has payments at first glance.